Deleted Fund , Added Transactions Fragment

+++ app/src/main/java/com/studycode/mymoneytracker/db/models/Budget.kt
@@ -8,10 +8,11 @@ class Budget(
     @PrimaryKey(autoGenerate = true)
     var id: Int?,
     val category: String,
-    var amount:Double?
+    var amount: Float?
 ) {
     constructor(
         category: String,
-        amount: Double?
-    ) : this(null, category,amount)
+        amount: Float?
+
+    ) : this(null, category, amount)
 }
